#0: fix ttnn package for profiler regression   : avoid circular import   : fix
